requirements.txt
discord.py>=2.3.2
httpx>=0.27.0
icalendar>=5.0.12
python-dotenv>=1.0.1
pytz>=2024.1
